# docs/openapi.yaml
...
components:
  schemas:
...
    BillingSettings:
      description: 'The community''s billing configuration.


        ``defaultMaintenanceAmount`` is null until an admin sets it. It has no home

        in the frontend or the ERD -- the amount exists today only as a literal

        ``4250`` inside an approval handler -- so there is nothing to migrate from

        and nothing to default to.'
      properties:
        autoBillingDay:
          default: 1
          title: Autobillingday
          type: integer
        autoBillingEnabled:
          default: false
          title: Autobillingenabled
          type: boolean
        communityId:
          title: Communityid
          type: string
        currency:
          title: Currency
          type: string
        defaultMaintenanceAmount:
          anyOf:
          - type: number
          - type: 'null'
          title: Defaultmaintenanceamount
        defaultTaxPercent:
          title: Defaulttaxpercent
          type: number
        invoiceNumberPrefix:
          title: Invoicenumberprefix
          type: string
        lateFeeAmount:
          anyOf:
          - type: number
          - type: 'null'
          title: Latefeeamount
        lateFeeEnabled:
          default: false
          title: Latefeeenabled
          type: boolean
        lateFeeGraceDays:
          default: 10
          title: Latefeegracedays
          type: integer
        lateFeePeriod:
          default: weekly
          title: Latefeeperiod
          type: string
        maintenanceDueDay:
          title: Maintenancedueday
          type: integer
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - communityId
      - currency
      - invoiceNumberPrefix
      - maintenanceDueDay
      - defaultTaxPercent
      - updatedAt
      title: BillingSettings
      type: object
...
    CreateInvoiceRequest:
      description: 'Issue one invoice against one flat.


        Either ``unitId`` or ``flat`` identifies the unit; ``flat`` is the code the

        rest of the app uses (``B-1204``) and is created on first reference, because

        the product has never had a flat-creation step.'
      properties:
        billingPeriodEnd:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Billingperiodend
        billingPeriodStart:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Billingperiodstart
        dueDate:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Duedate
        flat:
          anyOf:
          - type: string
          - type: 'null'
          title: Flat
        invoiceType:
          default: maintenance
          title: Invoicetype
          type: string
        issuedOn:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Issuedon
        lineItems:
          items:
            $ref: '#/components/schemas/LineItemInput'
          minItems: 1
          title: Lineitems
          type: array
        notes:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Notes
        taxPercent:
          anyOf:
          - exclusiveMaximum: 100.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Taxpercent
        title:
          maxLength: 160
          minLength: 1
          title: Title
          type: string
        unitId:
          anyOf:
          - type: string
          - type: 'null'
          title: Unitid
      required:
      - title
      - lineItems
      title: CreateInvoiceRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    InvoiceDetail:
      description: One invoice with everything hanging off it.
      properties:
        amount:
          title: Amount
          type: number
        amountPaid:
          title: Amountpaid
          type: number
        billPeriod:
          title: Billperiod
          type: string
        billingPeriodEnd:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Billingperiodend
        billingPeriodStart:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Billingperiodstart
        createdAt:
          format: date-time
          title: Createdat
          type: string
        currency:
          title: Currency
          type: string
        dueDate:
          format: date
          title: Duedate
          type: string
        flat:
          anyOf:
          - type: string
          - type: 'null'
          title: Flat
        id:
          title: Id
          type: string
        invoiceNumber:
          title: Invoicenumber
          type: string
        invoiceType:
          title: Invoicetype
          type: string
        isOverdue:
          title: Isoverdue
          type: boolean
        issuedOn:
          format: date
          title: Issuedon
          type: string
        lineItems:
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
          title: Lineitems
          type: array
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        outstanding:
          title: Outstanding
          type: number
        paidOn:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Paidon
        paymentMethod:
          anyOf:
          - type: string
          - type: 'null'
          title: Paymentmethod
        payments:
          items:
            $ref: '#/components/schemas/PaymentSummary'
          title: Payments
          type: array
        residentName:
          anyOf:
          - type: string
          - type: 'null'
          title: Residentname
        residentProfileId:
          anyOf:
          - type: string
          - type: 'null'
          title: Residentprofileid
        status:
          title: Status
          type: string
        statusDetail:
          title: Statusdetail
          type: string
        subtotal:
          title: Subtotal
          type: number
        tax:
          title: Tax
          type: number
        title:
          title: Title
          type: string
        tower:
          anyOf:
          - type: string
          - type: 'null'
          title: Tower
        unitId:
          anyOf:
          - type: string
          - type: 'null'
          title: Unitid
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
        userId:
          anyOf:
          - type: string
          - type: 'null'
          title: Userid
      required:
      - id
      - invoiceNumber
      - title
      - invoiceType
      - amount
      - subtotal
      - tax
      - outstanding
      - amountPaid
      - currency
      - status
      - statusDetail
      - isOverdue
      - dueDate
      - issuedOn
      - billPeriod
      - createdAt
      - updatedAt
      title: InvoiceDetail
      type: object
...
    Page_InvoiceSummary_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/InvoiceSummary'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[InvoiceSummary]
      type: object
...
    UpdateBillingSettingsRequest:
      description: 'Patch the billing configuration. Omitted fields are left unchanged.


        ``defaultMaintenanceAmount`` distinguishes omitted from null: sending null

        clears the rate, which stops billing runs until one is set again.'
      properties:
        autoBillingDay:
          anyOf:
          - maximum: 28.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Autobillingday
        autoBillingEnabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Autobillingenabled
        currency:
          anyOf:
          - maxLength: 3
            minLength: 3
            type: string
          - type: 'null'
          title: Currency
        defaultMaintenanceAmount:
          anyOf:
          - maximum: 10000000.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Defaultmaintenanceamount
        defaultTaxPercent:
          anyOf:
          - exclusiveMaximum: 100.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Defaulttaxpercent
        invoiceNumberPrefix:
          anyOf:
          - maxLength: 12
            minLength: 1
            pattern: ^[A-Za-z0-9-]+$
            type: string
          - type: 'null'
          title: Invoicenumberprefix
        lateFeeAmount:
          anyOf:
          - maximum: 1000000.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Latefeeamount
        lateFeeEnabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Latefeeenabled
        lateFeeGraceDays:
          anyOf:
          - maximum: 90.0
            minimum: 0.0
            type: integer
          - type: 'null'
          title: Latefeegracedays
        lateFeePeriod:
          anyOf:
          - maxLength: 16
            type: string
          - type: 'null'
          title: Latefeeperiod
        maintenanceDueDay:
          anyOf:
          - maximum: 28.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Maintenancedueday
      title: UpdateBillingSettingsRequest
      type: object
...
    VoidInvoiceRequest:
      description: Cancel an invoice. Refused once any payment has succeeded against it.
      properties:
        reason:
          anyOf:
          - maxLength: 500
            type: string
          - type: 'null'
          title: Reason
      title: VoidInvoiceRequest
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  summary: Backend for the HomeBandhu residential community platform.
  title: HomeBandhu API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/billing-settings:
    get:
      description: 'The community''s billing configuration.


        ``defaultMaintenanceAmount`` is null until an admin sets one. There is no

        screen for this today and no value to migrate: the maintenance amount exists

        in the product only as a literal ``4250`` inside an approval handler

        (agenda item 12).'
      operationId: get_billing_settings_api_v1_billing_settings_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingSettings'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get billing settings
      tags:
      - money
    put:
      description: 'Patch the billing configuration. Omitted fields are left unchanged.


        Sending ``defaultMaintenanceAmount: null`` clears the rate and stops billing

        runs until one is set again -- distinct from omitting the field, which leaves

        it as it was.


        ``invoiceNumberPrefix`` only affects numbers issued from now on. Invoices

        already issued keep the number they were given, because a number that changes

        is not an identifier.'
      operationId: update_billing_settings_api_v1_billing_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBillingSettingsRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingSettings'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update billing settings
      tags:
      - money
...
  /api/v1/invoices:
    get:
      description: 'The maintenance collections table, newest due date first.


        ``amount`` is what the flat was billed, not what it still owes -- the column

        is headed "Amount". ``outstanding`` carries the balance, and

        ``GET /invoices/summary`` is the authority on the totals: deriving them by

        summing this page gives the total of a page, not of the community.'
      operationId: list_invoices_api_v1_invoices_get
      parameters:
      - description: Matches the invoice title, its number, the flat code and the current resident's name
          -- the fields the collections search box covers.
        in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          description: Matches the invoice title, its number, the flat code and the current resident's
            name -- the fields the collections search box covers.
          title: Q
      - description: Paid | Unpaid | Void | All. A partially paid invoice is Unpaid.
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Paid | Unpaid | Void | All. A partially paid invoice is Unpaid.
          title: Status
      - in: query
        name: unitId
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Unitid
      - description: maintenance | amenity | penalty | misc
        in: query
        name: invoiceType
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: maintenance | amenity | penalty | misc
          title: Invoicetype
      - in: query
        name: overdueOnly
        required: false
        schema:
          default: false
          title: Overdueonly
          type: boolean
      - in: query
        name: issuedFrom
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Issuedfrom
      - in: query
        name: issuedTo
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Issuedto
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_InvoiceSummary_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List invoices
      tags:
      - money
    post:
      description: 'Issue one invoice against one flat.


        The invoice, its line items and its number are written in one transaction,

        and the totals are computed from the lines -- so a header that disagrees with

        its own contents cannot be submitted.


        The flat is identified by ``unitId`` or by ``flat`` (``B-1204``), and is

        created on first reference. Returns 409 when the lines total zero or the due

        date precedes the issue date.'
      operationId: create_invoice_api_v1_invoices_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Issue an invoice
      tags:
      - money
  /api/v1/invoices/summary:
    get:
      description: 'The three tiles at the top of the Maintenance screen.


        Aggregated by Postgres over every invoice in the community, so the figures do

        not change when the list is paged. ``totalOutstanding`` sums the outstanding

        *balances*, so a partially paid invoice contributes only what is still owed.


        A community with no invoices reports zeros with HTTP 200, never a 404.'
      operationId: get_collection_summary_api_v1_invoices_summary_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionSummaryDetail'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Collection totals
      tags:
      - money
  /api/v1/invoices/{invoice_id}:
    get:
      description: One invoice with its line items and every payment recorded against it.
      operationId: get_invoice_api_v1_invoices__invoice_id__get
      parameters:
      - in: path
        name: invoice_id
        required: true
        schema:
          title: Invoice Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get an invoice
      tags:
      - money
  /api/v1/invoices/{invoice_id}/payments:
    post:
      description: 'Record money received against an invoice and return the settled invoice.


        **Idempotent on ``reference``.** Sending the same reference twice returns the

        payment already recorded rather than crediting the invoice again, so a

        retried gateway webhook or a double-tapped Pay button cannot settle a bill

        twice.


        Returns 409 for an amount above the outstanding balance -- an overpayment is

        refused rather than clamped, because clamping accepts money and then loses

        it. Also 409 once the invoice has been voided.'
      operationId: record_payment_api_v1_invoices__invoice_id__payments_post
      parameters:
      - in: path
        name: invoice_id
        required: true
        schema:
          title: Invoice Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPaymentRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Record a payment
      tags:
      - money
  /api/v1/invoices/{invoice_id}/void:
    post:
      description: 'Cancel an invoice.


        **There is no DELETE for money.** The invoice, its lines and its number stay

        in place and are marked void -- an invoice number that vanishes is a gap

        somebody has to account for later.


        Returns 409 once any payment has succeeded against it: cancelling a bill

        somebody has already paid would strand their money against nothing.'
      operationId: void_invoice_api_v1_invoices__invoice_id__void_post
      parameters:
      - in: path
        name: invoice_id
        required: true
        schema:
          title: Invoice Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidInvoiceRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Void an invoice
      tags:
      - money
  /api/v1/maintenance-runs:
    post:
      description: 'Issue one maintenance invoice per **occupied** flat for a billing period.


        **Safe to repeat.** A second run for the same period reports every flat as

        skipped and bills nobody; the guard is a partial unique index, so it holds

        even if two admins click at the same moment.


        The amount comes from ``billingSettings.defaultMaintenanceAmount`` unless

        overridden here. With neither set the call returns **409** rather than

        falling back to a number nobody chose -- the frontend''s hardcoded 4250 is a

        demo value and adopting it would bill a real community by accident.


        Vacant flats are not billed, because nothing in the product records who owns

        an empty one (DECISIONS_NEEDED A14).'
      operationId: run_maintenance_billing_api_v1_maintenance_runs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRunRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRunResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Run maintenance billing
      tags:
      - money
...
  /api/v1/payments:
    get:
      description: 'The collection log the Maintenance screen''s subtitle promises.


        Every payment received, newest first, across all invoices -- which is a

        different question from "what does this flat owe" and needs its own list.'
      operationId: list_payments_api_v1_payments_get
      parameters:
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Q
      - description: UPI | Credit Card | Net Banking | Cash | Cheque | Bank Transfer
        in: query
        name: method
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: UPI | Credit Card | Net Banking | Cash | Cheque | Bank Transfer
          title: Method
      - in: query
        name: invoiceId
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Invoiceid
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_PaymentSummary_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List payments
      tags:
      - money
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Deployed
  url: https://{host}
  variables:
    host:
      default: api.homebandhu.example
      description: API host
...